Change Python bindings method for overriding virtual methods of C++ classes
To override the C++ virtual method 'DoStart', we new define a method in Python called 'DoStart', whilst before we would define '_DoStart'.  This change is to align pybindgen generated bindings with the other python bindings tools out there, such as boost.python, swig, and sip.  It is also simpler to understand for users.